refactor: 툴팁 컴포넌트 분리
