--- src/atoms/availableGuideAtoms.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const availableGuideState = atom<boolean>({
-  key: 'availableGuide',
-  default: true,
-});
