+++ src/atoms/isTooltipShownAtoms.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isTooltipShownState = atom<boolean>({
+  key: 'isTooltipShown',
+  default: true,
+});
